requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dateutils==0.6.12
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.78.0
fastcore==1.3.21
fastprogress==1.0.3
fastscript==1.0.0
gitdb==4.0.10
GitPython==3.1.27
giturlparse==0.10.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.0.0
loguru==0.6.0
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
orjson==3.8.6
outcome==1.2.0
packaging==23.0
Pillow==8.4.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymemri==0.0.45
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.19.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.17.6
uvloop==0.17.0
watchgod==0.8.2
websockets==10.4
wsproto==1.2.0
